requirements.txt
# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
urllib3>=2.0.0

# Data analysis
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Visualization
plotly>=5.14.0
streamlit>=1.28.0

# Scheduling
schedule>=1.2.0

# Additional utilities
python-dateutil>=2.8.0

# Phase 1 Improvements: Data Quality & Alerts
pydantic>=2.0.0
great-expectations>=0.18.0
sendgrid>=6.10.0
python-dotenv>=1.0.0

# Azure Blob Storage & DuckDB (for GitHub Actions sync)
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
duckdb>=0.9.0
pyarrow>=14.0.0

# Slack notifications (optional)
slack-sdk>=3.21.0

# Selenium for web scraping
selenium>=4.15.0
webdriver-manager>=4.0.0

# Phase 1 Improvements: PDF Parsing
camelot-py[cv]>=0.11.0
PyPDF2>=3.0.0
tabula-py>=2.7.0

# Phase 2 Improvements: Machine Learning (Future)
# scikit-learn>=1.3.0
# xgboost>=2.0.0
# statsmodels>=0.14.0
# shap>=0.43.0

# Phase 3 Improvements: Database & API (Future)
# sqlalchemy>=2.0.0
# psycopg2-binary>=2.9.0
# alembic>=1.12.0
# fastapi>=0.104.0
# uvicorn>=0.24.0

# Phase 5 Improvements: Advanced Analytics (Future)
# scipy>=1.11.0
# networkx>=3.1
# pyvis>=0.3.2
